Throw warning for mismatched Set types for all concrete implementations of AbstractSetFilter
