+++ src/main/java/com/hubspot/jinjava/lib/filter/AbstractSetFilter.java
@@ -1,6 +1,7 @@
 package com.hubspot.jinjava.lib.filter;
 
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
+import com.hubspot.jinjava.interpret.TemplateError;
 import com.hubspot.jinjava.interpret.TemplateSyntaxException;
 import com.hubspot.jinjava.lib.fn.TypeFunction;
 import com.hubspot.jinjava.util.ForLoop;
@@ -31,7 +32,41 @@ public abstract class AbstractSetFilter implements AdvancedFilter {
     return result;
   }
 
-  protected String getTypeOfSetElements(Set<Object> set) {
+  protected void attachMismatchedTypesWarning(
+    JinjavaInterpreter interpreter,
+    Set<Object> varSet,
+    Set<Object> argSet
+  ) {
+    boolean hasAtLeastOneSetEmpty = varSet.isEmpty() || argSet.isEmpty();
+    if (hasAtLeastOneSetEmpty) {
+      return;
+    }
+
+    boolean areMatchedElementTypes = getTypeOfSetElements(varSet)
+      .equals(getTypeOfSetElements(argSet));
+    if (areMatchedElementTypes) {
+      return;
+    }
+
+    interpreter.addError(
+      new TemplateError(
+        TemplateError.ErrorType.WARNING,
+        TemplateError.ErrorReason.OTHER,
+        TemplateError.ErrorItem.FILTER,
+        String.format(
+          "Mismatched Types: input set has elements of type '%s' but arg set has elements of type '%s'. Use |map filter to convert sets to the same type for filter to work correctly.",
+          getTypeOfSetElements(varSet),
+          getTypeOfSetElements(argSet)
+        ),
+        "list",
+        interpreter.getLineNumber(),
+        -1,
+        null
+      )
+    );
+  }
+
+  private String getTypeOfSetElements(Set<Object> set) {
     if (set.isEmpty()) {
       return "null";
     }
